refactor-edit-tab * Split every tile of EditTab into its own component * Move "Panel"-components into the tile directories * Use css-modules for styling
