+++ src/containers/Admin/EditTab/BikeTile/BikePanel/BikePanel.scss
