+++ src/containers/Admin/EditTab/CustomTile/CustomTile.module.scss
@@ -1,0 +1,10 @@
+.TooltipContent {
+    color: var(--tavla-box-background-color);
+    margin: 0;
+    max-width: 14rem;
+    white-space: normal;
+}
+
+.Icon {
+    margin-left: 0.5rem;
+}
