+++ src/containers/Admin/EditTab/CustomTile/CustomTilePanel/CustomTilePanel.scss
@@ -1,0 +1,48 @@
+@use '../../../../../assets/styles/Variables.scss';
+
+.custom-tile-panel {
+    .eds-button {
+        &.eds-button--variant-primary {
+            background: var(--floating-button-primary-background);
+            color: var(--floating-button-primary-font-color);
+            border: none;
+        }
+
+        &:hover {
+            filter: brightness(0.95);
+        }
+    }
+
+    &__add-button {
+        width: 100%;
+        margin-top: 1rem;
+        margin-bottom: 2rem;
+    }
+
+    &__row {
+        display: flex;
+        align-items: center;
+        height: 3rem;
+        justify-content: space-between;
+        font-size: Variables.$font-sizes-large;
+        font-weight: 600;
+        -webkit-font-smoothing: antialiased;
+
+        &:not(:last-child) {
+            border-bottom: 0.125rem solid var(--tavla-border-color);
+        }
+    }
+
+    &__icons {
+        display: flex;
+        align-items: center;
+
+        & > * {
+            cursor: pointer;
+        }
+
+        & > *:not(:last-child) {
+            margin-right: 1rem;
+        }
+    }
+}
