+++ src/containers/Admin/EditTab/EditTab.module.scss
@@ -1,0 +1,110 @@
+.TilesHeading {
+    margin: 2rem 2rem 0 2rem;
+    font-size: 1.3rem;
+    opacity: 0.75;
+}
+
+.TilesContainer {
+    display: flex;
+    flex-direction: row;
+    align-items: flex-start;
+    padding-left: 2rem;
+    padding-right: 2rem;
+    margin-top: 2rem;
+    gap: 3rem;
+}
+
+.EditTab :global {
+    /* Chrome, Safari, Edge, Opera */
+    .eds-form-control::-webkit-outer-spin-button,
+    .eds-form-control::-webkit-inner-spin-button {
+        -webkit-appearance: none;
+        margin: 0;
+    }
+
+    /* Firefox */
+    .eds-form-control[type='number'] {
+        -moz-appearance: textfield;
+    }
+
+    .eds-checkbox-icon {
+        background-color: var(--tavla-checkbox-color);
+    }
+
+    .eds-checkbox__icon,
+    .eds-contrast .eds-checkbox__icon {
+        border: 0.125rem solid var(--tavla-checkbox-color);
+        background-color: var(--tavla-checkbox-color);
+    }
+
+    .eds-checkbox__container input:checked + .eds-checkbox__icon,
+    .eds-checkbox__container input:indeterminate + .eds-checkbox__icon {
+        background-color: var(--tavla-checkbox-color);
+    }
+
+    .eds-checkbox__container:hover input + .eds-checkbox__icon,
+    .eds-contrast .eds-checkbox__container:hover input + .eds-checkbox__icon {
+        border-color: var(--tavla-checkbox-hover-color);
+        background-color: var(--tavla-checkbox-color);
+    }
+
+    .eds-form-control-wrapper {
+        border-color: transparent;
+        background-color: var(--tavla-input-background-color);
+
+        &:hover {
+            background-color: var(--tavla-input-background-color);
+            border-color: var(--tavla-input-border-hover-color);
+        }
+
+        .eds-input-group__label {
+            color: var(--tavla-input-placeholder-color);
+        }
+    }
+
+    .eds-dropdown__input {
+        color: var(--tavla-input-font-color);
+    }
+
+    .eds-dropdown__input::placeholder {
+        color: var(--tavla-input-font-color);
+    }
+
+    .eds-form-control__prepend svg,
+    .eds-form-control__append svg {
+        color: var(--tavla-input-icon-color);
+    }
+
+    .eds-grid {
+        &__item {
+            height: fit-content;
+
+            h2 {
+                margin: 0;
+                color: var(--tavla-font-color);
+                overflow: hidden;
+                white-space: nowrap;
+                text-overflow: ellipsis;
+            }
+        }
+    }
+
+    @media (max-width: 996px) {
+        padding: 0;
+
+        .eds-h3 {
+            margin: 1rem 0;
+        }
+    }
+
+    .eds-dropdown__input[value=''] + .eds-form-control__append {
+        display: none;
+    }
+}
+
+@media (min-width: 996px) {
+    .EditTab {
+        margin-left: -2rem;
+        margin-right: -2rem;
+    }
+}
