--- src/containers/Admin/EditTab/EditTab.scss
@@ -1,235 +1,0 @@
-@use '../../../assets/styles/Variables.scss';
-
-.edit-tab {
-    &__tiles-heading {
-        margin: 2rem 2rem 0 2rem;
-        font-size: 1.3rem;
-        opacity: 0.75;
-    }
-
-    &__tiles-container {
-        display: flex;
-        flex-direction: row;
-        align-items: flex-start;
-        padding-left: 2rem;
-        padding-right: 2rem;
-        margin-top: 2rem;
-        gap: 3rem;
-    }
-
-    .eds-contrast {
-        background: inherit;
-    }
-
-    .tooltip-container {
-        font-weight: 500;
-        max-width: 15rem;
-        padding: 0.5rem;
-
-        h4,
-        p {
-            color: var(--tavla-box-background-color);
-        }
-    }
-
-    .tooltip-container-weather {
-        color: var(--tavla-box-background-color);
-        margin: 0;
-    }
-
-    /* Chrome, Safari, Edge, Opera */
-    .eds-form-control::-webkit-outer-spin-button,
-    .eds-form-control::-webkit-inner-spin-button {
-        -webkit-appearance: none;
-        margin: 0;
-    }
-
-    /* Firefox */
-    .eds-form-control[type='number'] {
-        -moz-appearance: textfield;
-    }
-
-    .eds-checkbox-icon {
-        background-color: var(--tavla-checkbox-color);
-    }
-
-    .eds-checkbox__icon,
-    .eds-contrast .eds-checkbox__icon {
-        border: 0.125rem solid var(--tavla-checkbox-color);
-        background-color: var(--tavla-checkbox-color);
-    }
-
-    .eds-checkbox__container input:checked + .eds-checkbox__icon,
-    .eds-checkbox__container input:indeterminate + .eds-checkbox__icon {
-        background-color: var(--tavla-checkbox-color);
-    }
-
-    .eds-checkbox__container:hover input + .eds-checkbox__icon,
-    .eds-contrast .eds-checkbox__container:hover input + .eds-checkbox__icon {
-        border-color: var(--tavla-checkbox-hover-color);
-        background-color: var(--tavla-checkbox-color);
-    }
-
-    .eds-form-control-wrapper {
-        border-color: transparent;
-        background-color: var(--tavla-input-background-color);
-
-        &:hover {
-            background-color: var(--tavla-input-background-color);
-            border-color: var(--tavla-input-border-hover-color);
-        }
-
-        .eds-input-group__label {
-            color: var(--tavla-input-placeholder-color);
-        }
-    }
-
-    .eds-dropdown__input {
-        color: var(--tavla-input-font-color);
-    }
-
-    .eds-dropdown__input::placeholder {
-        color: var(--tavla-input-font-color);
-    }
-
-    .eds-form-control__prepend svg,
-    .eds-form-control__append svg {
-        color: var(--tavla-input-icon-color);
-    }
-
-    .eds-grid {
-        &__item {
-            height: fit-content;
-
-            h2 {
-                margin: 0;
-                color: var(--tavla-font-color);
-                overflow: hidden;
-                white-space: nowrap;
-                text-overflow: ellipsis;
-            }
-        }
-    }
-
-    &__header {
-        display: flex;
-        justify-content: space-between;
-        margin: 1rem 0 1rem 0;
-
-        .eds-switch {
-            padding: 0;
-        }
-
-        .eds-h2 {
-            margin: 0;
-            max-width: 75%;
-            text-overflow: ellipsis;
-            overflow: hidden;
-            white-space: nowrap;
-        }
-
-        &--details-in-view {
-            margin-top: 0;
-        }
-    }
-
-    &__expanding-text-field {
-        font-size: inherit;
-    }
-
-    &__expanding-text-field.eds-form-control-wrapper {
-        background-color: transparent;
-        font-weight: 400;
-        border-color: transparent;
-
-        &:focus-within {
-            --border-color: transparent;
-        }
-
-        &:hover {
-            background-color: transparent;
-        }
-
-        .eds-form-control,
-        .eds-form-control__append {
-            font-size: inherit;
-            white-space: nowrap;
-            text-align: center;
-            padding: 0;
-        }
-    }
-
-    &__input-wrapper {
-        display: inline-block;
-        position: relative;
-        bottom: Variables.$space-extra-small;
-    }
-
-    &__input-wrapper::after {
-        content: '';
-        display: block;
-        margin: 0 auto;
-        width: 7.5rem;
-        border-bottom: 5px dotted var(--tavla-input-dotted-border-color);
-    }
-
-    .distance-editor {
-        margin-left: 2rem;
-        width: auto;
-    }
-
-    .zoom-editor {
-        width: auto;
-    }
-
-    .stop-place-search {
-        flex-grow: 1;
-    }
-
-    @media (max-width: 996px) {
-        padding: 0;
-
-        &__expanding-text-field {
-            padding: 0;
-        }
-
-        &__input-wrapper {
-            padding: 0;
-        }
-
-        &__input-wrapper::after {
-            width: 5rem;
-            margin: 0;
-            margin-top: -1rem;
-            margin-left: 0.2rem;
-            border-bottom: 5px dotted var(--tavla-input-dotted-border-color);
-        }
-
-        .eds-h3 {
-            margin: 1rem 0;
-        }
-    }
-
-    .eds-dropdown__input[value=''] + .eds-form-control__append {
-        display: none;
-    }
-}
-
-@media (min-width: 500px) {
-    .edit-tab {
-        &__heading {
-            margin-left: 2rem;
-        }
-
-        &__input-wrapper {
-            width: 10rem;
-        }
-    }
-}
-
-@media (min-width: 996px) {
-    .edit-tab {
-        margin-left: -2rem;
-        margin-right: -2rem;
-    }
-}
