+++ src/containers/Admin/EditTab/MobileAppTile/MobileAppTile.module.scss
@@ -1,0 +1,10 @@
+.TooltipParagraph {
+    color: var(--tavla-box-background-color);
+    margin: 0;
+    max-width: 14rem;
+    white-space: normal;
+}
+
+.Icon {
+    margin-left: 0.5rem;
+}
