+++ src/containers/Admin/EditTab/RealtimeTile/RealtimeDataPanel/linesPanel.scss
