+++ src/containers/Admin/EditTab/RealtimeTile/RealtimeTile.tsx
@@ -1,0 +1,36 @@
+import React, { useCallback } from 'react'
+import { Paragraph } from '@entur/typography'
+import { EditTile } from '../EditTile/EditTile'
+import { useSettings } from '../../../../settings/SettingsProvider'
+import { RealtimeDataPanel } from './RealtimeDataPanel/RealTimeDataPanel'
+
+const RealtimeTile: React.FC = () => {
+    const [settings, setSettings] = useSettings()
+
+    const handleChange = useCallback(() => {
+        setSettings({
+            hideRealtimeData: !settings.hideRealtimeData,
+        })
+    }, [settings, setSettings])
+
+    return (
+        <EditTile
+            title="Sanntidsposisjoner"
+            onChange={handleChange}
+            checked={!settings.hideRealtimeData}
+        >
+            {!settings.hiddenModes.includes('kollektiv') ? (
+                <RealtimeDataPanel
+                    hiddenLines={settings.hiddenRealtimeDataLineRefs}
+                />
+            ) : (
+                <Paragraph>
+                    Kollektivdata er skrudd av. Skru det på ved å trykke på
+                    knappen øverst til høyre i kollektiv-ruten.
+                </Paragraph>
+            )}
+        </EditTile>
+    )
+}
+
+export { RealtimeTile }
