+++ src/containers/Admin/EditTab/ReloadTile/ReloadTile.module.scss
@@ -1,4 +1,4 @@
-.EditTile {
+.ReloadTile {
     margin-bottom: 3rem;
     padding: 2rem;
     background-color: var(--tavla-box-background-color);
@@ -19,8 +19,4 @@
     text-overflow: ellipsis;
     overflow: hidden;
     white-space: nowrap;
-}
-
-.Switch {
-    padding: 0;
 }
