+++ src/containers/Admin/EditTab/ScooterTile/ScooterPanel/ScooterPanel.mobility.graphql
