+++ src/containers/Admin/EditTab/ScooterTile/ScooterPanel/ScooterPanel.scss
