--- src/containers/Admin/EditTab/StopPlacePanel/StopPlacePanel.scss
@@ -1,103 +1,0 @@
-@use '../../../../assets/styles/Variables.scss';
-
-.stop-place-panel {
-    margin: 1rem 0 1rem 0;
-
-    .eds-paragraph {
-        color: var(--tavla-font-color);
-    }
-
-    &__header {
-        display: flex;
-        padding: 1rem 0;
-        align-items: center;
-    }
-
-    &__row {
-        display: flex;
-        flex-direction: row;
-        align-items: center;
-
-        &__empty {
-            padding: 1rem 1.5rem 1rem 0;
-        }
-
-        .eds-expandable-panel__trigger,
-        .eds-expandable-panel__trigger:hover {
-            background: none;
-            padding-left: 0;
-            box-shadow: none;
-        }
-
-        &__header {
-            display: flex;
-            flex-direction: row;
-            align-items: center;
-            flex-wrap: wrap;
-
-            .eds-switch {
-                margin-left: 1rem;
-            }
-
-            > span {
-                display: flex;
-                color: var(--tavla-font-color);
-                font-weight: 600;
-                -webkit-font-smoothing: antialiased;
-                align-items: center;
-                font-size: Variables.$font-sizes-large;
-            }
-        }
-
-        &__content {
-            display: flex;
-            flex-direction: column;
-        }
-
-        &__expandable {
-            flex: 1;
-
-            > .entur-expandable__content {
-                padding: 1rem 0;
-            }
-        }
-
-        &__item-button {
-            display: flex;
-            flex-direction: row;
-            align-items: center;
-        }
-    }
-
-    &__route-checkbox {
-        margin-bottom: 0.5rem;
-
-        > label {
-            color: var(--tavla-font-color);
-        }
-    }
-
-    &__route {
-        display: flex;
-        flex-direction: row;
-        align-items: center;
-        margin-bottom: 0.5rem;
-
-        .eds-checkbox__label {
-            color: var(--tavla-font-color);
-            display: flex;
-            align-items: center;
-            white-space: nowrap;
-
-            svg {
-                font-size: 1.5rem;
-                margin-right: 0.5rem;
-            }
-        }
-    }
-
-    &__header,
-    &__row:not(:last-child) {
-        border-bottom: 0.125rem solid var(--tavla-border-color);
-    }
-}
