--- src/containers/Admin/EditTab/StopPlaceSearch/StopPlaceSearch.module.scss
@@ -1,3 +1,0 @@
-.StopPlaceSearch {
-    width: 100%;
-}
