+++ src/containers/Admin/EditTab/StopPlaceTile/StopPlaceSearch/StopPlaceSearch.module.scss
@@ -1,0 +1,3 @@
+.StopPlaceSearch {
+    width: 100%;
+}
