+++ src/containers/Admin/EditTab/StopPlaceTile/StopPlaceTile.module.scss
@@ -1,0 +1,3 @@
+.DetailsInView {
+    margin-top: 0;
+}
