+++ src/containers/Admin/EditTab/WeatherTile/WeatherPanel/WeatherPanel.scss
